Update HCL dependency to fix panic bug
This is selecting a commit that has not yet been included in an HCL
release, in order to make this tool work better for folks who had
expressions of the sort that were tripping over the crashing bug.

This fixes #2. We acn pin this back to a real HCL release again once this
fix is included in one.